Updatecripts in install-dotnet document
The default value in the script (For .NET Fx 4.6.2) and the hyperlink
provided in the document before (.NET Fx 4.6.1) doesn't match each other,
which will lead to deploy failure unless corrected by user manually.

Update the default value in the script provided to use Fx 4.6.1 to match
the download hyperlink.